ver-0.2

+++ src/main/java/hr/klobucaric/webshop/address/Address.java
@@ -4,7 +4,6 @@ import hr.klobucaric.webshop.shopOrder.ShopOrder;
 import hr.klobucaric.webshop.user.User;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.Hibernate;
 
 import javax.persistence.*;
 import java.util.HashSet;
@@ -21,7 +20,7 @@ import java.util.Set;
 public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     public Address(String city, Integer postalCode, String addressLine, Integer streetNumber) {
         this.city = city;
@@ -44,9 +43,6 @@ public class Address {
     public void addUserAddress(User user, Boolean isDefault) {
         userAddresses.add( new UserAddress(user, this, isDefault));
     }
-
-    //todo https://www.baeldung.com/jpa-many-to-many
-    //todo boolean se mora promjeniti ostalima u nula dok se nekome dene na 1
 
 
     @Override
